task 6

+++ hello_world/6-size.c
@@ -6,10 +6,10 @@
  */
 int main(void)
 {
-printf("Size of char: %lu bytes\n", sizeof(char));
-printf("Size of int: %lu bytes\n", sizeof(int));
-printf("Size of long int: %lu bytes\n", sizeof(long int));
-printf("Size of long long int: %lu bytes\n", sizeof(double));
-printf("Size of float: %lu bytes\n", sizeof(float));
+  printf("Size of char: %u bytes\n",(int) sizeof(char));
+  printf("Size of int: %u bytes\n",(int) sizeof(int));
+  printf("Size of long int: %u bytes\n",(int) sizeof(long int));
+  printf("Size of long long int: %u bytes\n",(int) sizeof(double));
+  printf("Size of float: %u bytes\n",(int) sizeof(float));
 return 0;
 }
